Add support for extracting xattrs on unix
When extended file attributes are set on a file in the archive, try to
preserve them on extraction but do not fail when the user does not have
permission to set them or if they are not supported. This should prevent
existing usages from breaking while also enabling users to preserve
extended file attributes.